IOS-5636 Move on `ExpressCurrencyView`
