+++ Tangem/Common/UI/SendCurrencyView/SendCurrencyView.swift
@@ -9,10 +9,9 @@
 import SwiftUI
 
 struct SendCurrencyView: View {
-    private var viewModel: SendCurrencyViewModel
+    @ObservedObject private var viewModel: SendCurrencyViewModel
     @Binding private var decimalValue: DecimalNumberTextField.DecimalValue?
 
-    private let tokenIconSize = CGSize(width: 36, height: 36)
     private var didTapChangeCurrency: (() -> Void)?
     private var didTapMaxAmountAction: (() -> Void)?
 
@@ -22,76 +21,19 @@ struct SendCurrencyView: View {
     }
 
     var body: some View {
-        VStack(spacing: 8) {
-            headerLabels
-
-            mainContent
+        ExpressCurrencyView(viewModel: viewModel.expressCurrencyViewModel) {
+            SendDecimalNumberTextField(
+                decimalValue: $decimalValue,
+                maximumFractionDigits: viewModel.maximumFractionDigits,
+                font: Fonts.Regular.title1
+            )
+            .maximumFractionDigits(viewModel.maximumFractionDigits)
+            .didTapMaxAmount { didTapMaxAmountAction?() }
+            .simultaneousGesture(TapGesture().onEnded {
+                viewModel.textFieldDidTapped()
+            })
         }
-        .padding(.all, 14)
-        .background(Colors.Background.action)
-        .cornerRadius(14)
-    }
-
-    private var headerLabels: some View {
-        HStack(spacing: 0) {
-            switch viewModel.headerState {
-            case .header:
-                Text(Localization.swappingFromTitle)
-                    .style(Fonts.Regular.footnote, color: Colors.Text.secondary)
-            case .insufficientFunds:
-                Text(Localization.swappingInsufficientFunds)
-                    .style(Fonts.Regular.caption1, color: Colors.Text.warning)
-            }
-
-            Spacer()
-
-            switch viewModel.balance {
-            case .idle:
-                EmptyView()
-            case .loading:
-                SkeletonView()
-                    .frame(width: 100, height: 13)
-                    .cornerRadius(6)
-            case .formatted(let value):
-                SensitiveText(builder: Localization.commonBalance, sensitive: value)
-                    .style(Fonts.Regular.footnote, color: Colors.Text.tertiary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-    }
-
-    private var currencyContent: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            SendDecimalNumberTextField(decimalValue: $decimalValue, maximumFractionDigits: viewModel.maximumFractionDigits, font: Fonts.Regular.title1)
-                .maximumFractionDigits(viewModel.maximumFractionDigits)
-                .didTapMaxAmount { didTapMaxAmountAction?() }
-                .simultaneousGesture(TapGesture().onEnded {
-                    viewModel.textFieldDidTapped()
-                })
-
-            switch viewModel.fiatValue {
-            case .idle:
-                EmptyView()
-            case .loading:
-                SkeletonView()
-                    .frame(width: 50, height: 13)
-                    .cornerRadius(6)
-            case .formatted(let value):
-                Text(value)
-                    .style(Fonts.Regular.footnote, color: Colors.Text.tertiary)
-            }
-        }
-    }
-
-    private var mainContent: some View {
-        HStack(alignment: .top, spacing: 0) {
-            currencyContent
-
-            Spacer()
-
-            SwappingTokenIconView(state: viewModel.tokenIconState)
-                .onTap(viewModel.canChangeCurrency ? didTapChangeCurrency : nil)
-        }
+        .didTapChangeCurrency { didTapChangeCurrency?() }
     }
 }
 
@@ -110,30 +52,73 @@ extension SendCurrencyView: Setupable {
 struct SendCurrencyView_Preview: PreviewProvider {
     @State private static var decimalValue: DecimalNumberTextField.DecimalValue? = nil
 
-    static let viewModels: [SendCurrencyViewModel] = [
+    static let viewModels = [
         SendCurrencyViewModel(
-            balance: .loading,
-            fiatValue: .loading,
-            maximumFractionDigits: 8,
-            canChangeCurrency: true,
-            tokenIconState: .icon(
-                TokenIconInfoBuilder().build(
-                    from: .blockchain(.bitcoin(testnet: false)),
-                    isCustom: false
-                )
-            )
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .loading,
+                fiatAmountState: .loading,
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .blockchain(.ethereum(testnet: false)), isCustom: false)),
+                symbolState: .loaded(text: "ETH"),
+                canChangeCurrency: false
+            ),
+            maximumFractionDigits: 8
         ),
         SendCurrencyViewModel(
-            balance: .formatted("3043.75"),
-            fiatValue: .formatted("1000.71"),
-            maximumFractionDigits: 8,
-            canChangeCurrency: true,
-            tokenIconState: .icon(
-                TokenIconInfoBuilder().build(
-                    from: .blockchain(.bitcoin(testnet: false)),
-                    isCustom: false
-                )
-            )
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .formatted("0.0058"),
+                fiatAmountState: .loading,
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .blockchain(.cardano(extended: false)), isCustom: false)),
+                symbolState: .loaded(text: "ADA"),
+                canChangeCurrency: false
+            ),
+            maximumFractionDigits: 8
+        ),
+        SendCurrencyViewModel(
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .formatted("0.0058"),
+                fiatAmountState: .loading,
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .blockchain(.polygon(testnet: false)), isCustom: false)),
+                symbolState: .loaded(text: "MATIC"),
+                canChangeCurrency: true
+            ),
+            maximumFractionDigits: 8
+        ),
+        SendCurrencyViewModel(
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .formatted("0.0058"),
+                fiatAmountState: .loaded(text: "1100.46"),
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .blockchain(.polygon(testnet: false)), isCustom: false)),
+                symbolState: .loaded(text: "MATIC"),
+                canChangeCurrency: true
+            ),
+            maximumFractionDigits: 8
+        ),
+        SendCurrencyViewModel(
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .formatted("0.0058"),
+                fiatAmountState: .loaded(text: "2100.46 $"),
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .token(.tetherMock, .polygon(testnet: false)), isCustom: false)),
+                symbolState: .loaded(text: "USDT"),
+                canChangeCurrency: true
+            ),
+            maximumFractionDigits: 8
+        ),
+        SendCurrencyViewModel(
+            expressCurrencyViewModel: .init(
+                titleState: .text(Localization.swappingFromTitle),
+                balanceState: .formatted("0.0058"),
+                fiatAmountState: .loaded(text: "2100.46 $"),
+                priceChangePercent: "-24.3 %",
+                tokenIconState: .icon(TokenIconInfoBuilder().build(from: .token(.tetherMock, .polygon(testnet: false)), isCustom: false)),
+                symbolState: .loaded(text: "USDT"),
+                canChangeCurrency: true
+            ),
+            maximumFractionDigits: 8
         ),
     ]
 
@@ -142,8 +127,13 @@ struct SendCurrencyView_Preview: PreviewProvider {
             Colors.Background.secondary
 
             VStack {
-                ForEach(viewModels) {
-                    SendCurrencyView(viewModel: $0, decimalValue: $decimalValue)
+                ForEach(viewModels) { viewModel in
+                    GroupedSection(viewModel) { viewModel in
+                        SendCurrencyView(viewModel: viewModel, decimalValue: $decimalValue)
+                    }
+                    .interSectionPadding(12)
+                    .interItemSpacing(10)
+                    .verticalPadding(0)
                 }
             }
             .padding(.horizontal, 16)
