+++ Tangem/Common/UI/SendCurrencyView/SendCurrencyViewModel.swift
@@ -10,27 +10,15 @@ import Foundation
 import Combine
 
 class SendCurrencyViewModel: ObservableObject, Identifiable {
-    @Published private(set) var headerState: HeaderState = .header
-    @Published private(set) var fiatValue: State
-    @Published private(set) var canChangeCurrency: Bool
+    @Published private(set) var expressCurrencyViewModel: ExpressCurrencyViewModel
     @Published private(set) var maximumFractionDigits: Int
-    @Published private(set) var balance: State
-    @Published private(set) var tokenIconState: SwappingTokenIconView.State
-
-    private var walletDidChangeSubscription: AnyCancellable?
 
     init(
-        balance: State = .idle,
-        fiatValue: State = .idle,
-        maximumFractionDigits: Int,
-        canChangeCurrency: Bool,
-        tokenIconState: SwappingTokenIconView.State
+        expressCurrencyViewModel: ExpressCurrencyViewModel,
+        maximumFractionDigits: Int
     ) {
-        self.balance = balance
-        self.fiatValue = fiatValue
+        self.expressCurrencyViewModel = expressCurrencyViewModel
         self.maximumFractionDigits = maximumFractionDigits
-        self.canChangeCurrency = canChangeCurrency
-        self.tokenIconState = tokenIconState
     }
 
     func textFieldDidTapped() {
@@ -38,75 +26,11 @@ class SendCurrencyViewModel: ObservableObject, Identifiable {
     }
 
     func update(wallet: WalletModel, initialWalletId: Int) {
-        canChangeCurrency = wallet.id != initialWalletId
+        expressCurrencyViewModel.update(wallet: .loaded(wallet), initialWalletId: initialWalletId)
         maximumFractionDigits = wallet.decimalCount
-        tokenIconState = .icon(
-            TokenIconInfoBuilder().build(from: wallet.tokenItem, isCustom: wallet.isCustom)
-//            symbol: wallet.tokenItem.currencySymbol
-        )
-
-        walletDidChangeSubscription = wallet.walletDidChangePublisher.sink { [weak self] state in
-            switch state {
-            case .created, .loading:
-                self?.balance = .loading
-            case .idle:
-                let formatted = wallet.balanceValue.map { BalanceFormatter().formatDecimal($0) }
-                self?.balance = .formatted(formatted ?? BalanceFormatter.defaultEmptyBalanceString)
-            case .noAccount, .failed, .noDerivation:
-                self?.balance = .formatted(BalanceFormatter.defaultEmptyBalanceString)
-            }
-        }
     }
 
     func updateSendFiatValue(amount: Decimal?, tokenItem: TokenItem) {
-        guard let amount = amount else {
-            update(fiatValue: .formatted(BalanceFormatter().formatFiatBalance(0)))
-            return
-        }
-
-        guard let currencyId = tokenItem.currencyId else {
-            update(fiatValue: .formatted(BalanceFormatter.defaultEmptyBalanceString))
-            return
-        }
-
-        if let fiatValue = BalanceConverter().convertToFiat(value: amount, from: currencyId) {
-            let formatted = BalanceFormatter().formatFiatBalance(fiatValue)
-            update(fiatValue: .formatted(formatted))
-            return
-        }
-
-        fiatValue = .loading
-
-        runTask(in: self) { [currencyId] viewModel in
-            let fiatValue = try await BalanceConverter().convertToFiat(value: amount, from: currencyId)
-            let formatted = BalanceFormatter().formatFiatBalance(fiatValue)
-            try Task.checkCancellation()
-
-            await runOnMain {
-                viewModel.update(fiatValue: .formatted(formatted))
-            }
-        }
-    }
-
-    func update(fiatValue: State) {
-        self.fiatValue = fiatValue
-    }
-
-    func update(headerState: HeaderState) {
-        // TODO: Add animation for funds
-        self.headerState = headerState
-    }
-}
-
-extension SendCurrencyViewModel {
-    enum HeaderState {
-        case header
-        case insufficientFunds
-    }
-
-    enum State: Hashable {
-        case idle
-        case loading
-        case formatted(_ value: String)
+        expressCurrencyViewModel.updateReceiveCurrencyValue(expectAmount: amount, tokenItem: tokenItem)
     }
 }
